Django/requirements.txt
# Django
Django==4.2.6
django-cors-headers==4.1.0
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-debug-toolbar==4.2.0
django-cleanup==8.0.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.23
drf-yasg==1.21.7

# Engine
asgiref==3.7.2
python-dotenv==1.0.0
pillow==10.1.0
sqlparse==0.4.4
uritemplate==4.1.1
attrs==23.1.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
typing_extensions==4.8.0
sorl-thumbnail==12.10.0
setuptools==69.0.2

# Codestyle
flake8==6.1.0